Refactored ps.py, main.py, and the server README
