+++ server/README.md
@@ -2,7 +2,7 @@
 
 ## Setup
 
-```bash
+```console
 virtualenv -p python3.7 venv
 source venv/bin/activate
 pip install -r requirements.txt
@@ -15,7 +15,7 @@ on a platform using an ingress gateway such as Knative / Cloud Run).
 
 Replace `123456` by your Photoshop remote connection password.
 
-```python
+```console
 python src/main.py \
     --basnet_service_ip="http://X.X.X.X" \
     --basnet_service_host="basnet-http.default.example.com" \
