+++ server/src/main.py
@@ -79,7 +79,7 @@ def save():
     with open('cut_mask.png', 'wb') as f:
         f.write(res.content)
         # shutil.copyfileobj(res.raw, f)
-    
+
     logging.info(' > opening mask...')
     mask = Image.open('cut_mask.png').convert("L")
 
@@ -165,7 +165,9 @@ def paste():
         logging.info(' > sending to photoshop...')
         name = datetime.today().strftime('%Y-%m-%d-%H:%M:%S')
         img_path = os.path.join(os.getcwd(), 'cut_current.png')
-        ps.paste(img_path, name, x, y, password=args.photoshop_password)
+        if not ps.paste(img_path, name, x, y, password=args.photoshop_password):
+            logging.error('Something went wrong with the photoshop script execution')
+            return
     else:
         logging.info('screen not found')
