+++ server/src/ps.py
@@ -1,4 +1,5 @@
 from photoshop import PhotoshopConnection
+from os.path import basename
 
 # TODO: This offset should be detected by getTopLeft() but the new version
 # of Photoshop doesn't seem to support executeActionGet so we put it
@@ -10,46 +11,16 @@ DOC_WIDTH = 2121
 DOC_HEIGHT = 1280
 
 def paste(filename, name, x, y, password='123456'):
-    
+
     # There seem to be a bug on Windows where the path must be using unix separators.
     # https://github.com/cyrildiagne/ar-cutpaste/issues/5
     filename = filename.replace('\\', '/')
-    
-    with PhotoshopConnection(password=password) as conn:
-        script = """
-        function pasteImage(filename, layerName, x, y) {
-            var fileRef = new File(filename);
-            var doc = app.activeDocument;
-            
-            var currentLayer = doc.artLayers.add();
-            var curr_file = app.open(fileRef);
-            curr_file.selection.selectAll();
-            curr_file.selection.copy();
-            curr_file.close();
 
-            doc.paste();
-            doc.activeLayer.name = layerName;
-            doc.activeLayer.translate(x, y);
-            try {
-                doc.activeLayer.move(doc.layers[doc.layers.length - 1], ElementPlacement.PLACEBEFORE);
-            } catch(e) {
-                alert(e);
-            }
-        }
-        function getTopLeft() {
-            try {
-                var r = new ActionReference();
-                var bounds = executeActionGet(r)
-                    .getObjectValue(stringIDToTypeID("viewInfo"))
-                    .getObjectValue(stringIDToTypeID("activeView"))
-                    .getObjectValue(stringIDToTypeID("globalBounds"));
-                alert(t)
-            } catch (e) {
-                alert(e);
-            }
-        }
-        """
+    with PhotoshopConnection(password=password) as conn:
+        script = open(basename(__file__) + '/script.js', 'r').read()
         x -= DOC_WIDTH * 0.5 + DOC_OFFSET_X
         y -= DOC_HEIGHT * 0.5 + DOC_OFFSET_Y
         script += f'pasteImage("{filename}", "{name}", {x}, {y})'
-        conn.execute(script)
+        result = conn.execute(script)
+        if result.status != 0:
+            return False
